fix(wasm): resolve geometry-job types legacy-aware in all three pre-pass paths (#3187)
Seven sites in the browser pre-pass discovery paths resolved a job's type with
a bare `IfcType::from_str`, which knows only the current schema. A keyword that
a newer schema dropped came back `IfcType::Unknown(crc32)`, while the gate that
admitted the job (`has_geometry_by_name`) resolved it legacy-aware. The label
and the gate disagreed about the same keyword.

SCOPE, STATED PRECISELY. This is a consistency fix, not a user-visible one, and
an earlier draft of this message claimed otherwise. The label these sites
compute does not reach the wire today: `emit_jobs_chunk` writes three slots per
job (`id`, `start`, `end`) and binds the type to `_`, and the worker re-derives
it from the record via `legacy_aware_ifc_type_from_record` (`batch.rs:430`),
which is where #3179 was actually fixed. The wrong labels were dead data. What
changes is that the field stops being an `Unknown` waiting to surface if any
consumer starts reading it, and the two branches of one walk stop contradicting
each other.

Reachable rather than theoretical: enumerating `LEGACY_ENTITY_NAMES` against
`has_geometry_by_name` gives 22 keywords that reach these branches and are
`Unknown` under the bare resolver:

  IFCBEAMSTANDARDCASE IFCCOLUMNSTANDARDCASE IFCMEMBERSTANDARDCASE
  IFCPLATESTANDARDCASE IFCSLABSTANDARDCASE IFCDOORSTANDARDCASE
  IFCWINDOWSTANDARDCASE IFCOPENINGSTANDARDCASE IFCSLABELEMENTEDCASE
  IFCWALLELEMENTEDCASE IFCPROXY IFCBUILDINGELEMENT IFCEQUIPMENTELEMENT
  IFCELECTRICDISTRIBUTIONPOINT IFCELECTRICALELEMENT IFCCHAMFEREDGEFEATURE
  IFCROUNDEDEDGEFEATURE IFCSTRUCTURALLINEARACTIONVARYING
  IFCSTRUCTURALPLANARACTIONVARYING IFCSOLIDSTRATUM IFCVOIDSTRATUM
  IFCWATERSTRATUM

These are not all IFC2X3, which an earlier draft got wrong, and the first
correction of that was itself imprecise. Measured by exact name against the
generated per-version tables, the 22 split:

  10  IFC4-only, removed in IFC4X3   the StandardCase/ElementedCase family
   7  IFC2X3-only                    IFCEQUIPMENTELEMENT, IFCELECTRICDISTRIBUTIONPOINT,
                                     IFCELECTRICALELEMENT, the two edge-feature leaves,
                                     the two ...ActionVarying leaves
   2  in BOTH                        IFCPROXY, IFCBUILDINGELEMENT
   3  in NEITHER                     the IFC4X3 stratum leaves

The earlier draft named IFCPROXY as an example of "genuinely IFC2X3"; it is in
both tables, so it was the one keyword in the set that could not make that
point. IFCEQUIPMENTELEMENT is IFC2X3-only and is the right example.

The seventh site was missed on the first pass and is the reason this is worth
doing now: `gpu_meshes/prepass_discovery.rs`, the sharded column-discovery walk.
Its sibling branch eight lines above already resolves through the legacy-aware
`type_product_ifc_type`, under a comment stating that the label can never
disagree with the gate that admitted it. The geometry-job branch below it was
the standing counterexample to that comment.

Both new tests fail with their own site reverted, with the diagnosis attached:
`IFCBEAMSTANDARDCASE was scheduled as Unknown(2906023776)` for the serial path,
`buffered_jobs = [(30, 412, 482, Unknown(2906023776))]` for the sharded one.
Each carries one real anti-vacuity guard -- that the keyword is still one the
BARE resolver gets wrong -- because if the generated enum ever learns it, the
test would otherwise pass with the fix removed.

Two files were on, or would have joined, the module-size ratchet, so their test
modules moved into `_tests.rs` siblings rather than growing inline, matching the
`prepass_orphan_type_tests.rs` convention already in the tree. `prepass.rs`
drops 430 -> 368, under the 400-line house rule, so its allowlist row is deleted
rather than lowered and the ratchet digest re-pinned; the justification comment
that row carried is deleted with it. `prepass_discovery.rs` drops to 162 and so
needs no row at all.
